Fixed some of my earlier fixes
Updated logging
Added commented-out test data for fail/splatfest error pages
Added logic to handle the data we do get if we don't get curr, next, and last maps

Cloudpebble's emulator isn't working atm, so I've only been able to test Aplite on my watch, and it's buggy as hell right now. I don't know if that's because of the updates or if it's because Cloudpebble is wonky.