Ongoing fixes, huge revamp
Figured out the proper syntax for changing the values of visual elements on the fly, implemented that.
The previous implementation had separate visual elements for each piece of text and each image. Now I've updated it to use minimal windows and static elements for each window, and I'm just updating the values instead of adding and removing elements.

It looks like the Aplite bug is a memory leak, which I hoped the above work would fix, but it remains a problem. I may have to just rewrite this in C if I want to clear that up.