Fixed the debug menu in Aplite
Not entirely sure why, but the colors I was using for the menu were inverted from the defaults. My background colors were being applied, but not my text colors, so in Aplite, i was getting black on black and white on white. I removed the custom values and the defaults are working perfectly as intended.