v2.2 release candidate!
FINALLY OK
Fixed images in Basalt and Aplite, it looks like the tags were borking everything.
Refactored a lot of splatmonitor.js to handle the changes to the Splatoon.ink API response data.
Removed some cruft and some old comments and duplicated code.
Added new debug case for the new Splatfest flag in the API response.
Added a lot of documentation for the API response structure and for a lot of the control flow and data structures.